+++ service/history/queues/action_slice_count.go
@@ -77,8 +77,8 @@ func (a *actionSliceCount) Run(readerGroup *ReaderGroup) {
 	// have to compact (force merge) slices to reduce slice count
 	preferredSliceCount := int(float64(a.attributes.CriticalSliceCount) * targetLoadFactor)
 
-	isDefaultReader := func(readerID int32) bool { return readerID == DefaultReaderId }
-	isNotDefaultReader := func(readerID int32) bool { return !isDefaultReader(readerID) }
+	isDefaultReader := func(readerID int64) bool { return readerID == DefaultReaderId }
+	isNotDefaultReader := func(readerID int64) bool { return !isDefaultReader(readerID) }
 	isUniversalPredicate := func(s Slice) bool { return tasks.IsUniverisalPredicate(s.Scope().Predicate) }
 	isNotUniversalPredicate := func(s Slice) bool { return !isUniversalPredicate(s) }
 
@@ -132,8 +132,8 @@ func (a *actionSliceCount) Run(readerGroup *ReaderGroup) {
 }
 
 func (a *actionSliceCount) findAndCompactCandidates(
-	readers map[int32]Reader,
-	readerPredicate func(int32) bool,
+	readers map[int64]Reader,
+	readerPredicate func(int64) bool,
 	slicePredicate SlicePredicate,
 	targetSliceCount int,
 ) bool {
