+++ service/history/queues/alerts.go
@@ -47,7 +47,7 @@ type (
 	}
 
 	AlertAttributesReaderStuck struct {
-		ReaderID         int32
+		ReaderID         int64
 		CurrentWatermark tasks.Key
 	}
